Start working on smart duration change
I want to make the timer calculate, after you change the duration
of a tomato or a break, where you would be if you used it with the new
setting from the beginning.

For example, you're doing a 45-minute-long tomato. 13 minutes left
before the timer goes off. And suddenly, you change the duration to
15 minutes. You did 32 minutes already. Won't they count anymore?
To prevent the loss of progress, I want the timer to count this time
and put you in the point where you would be if you used the tomato with
15-minute-long tomatoes from the beginning.

I will return to this task tomorrow.