Add functional add-task-btn, almost fix todo.js
As soon as I made the Add Task button functional, I realized that I
needed to change the logic of the task list in such a way that would've
allowed users to create and work with same-name tasks.

I refactored the todo.js module. Most of the methods now work with
same-name tasks. The only method left to refactor is .clear().

The rest of the modules currently are remaining to be refactored. Will
do soon.